refactor: move legacy catalog resolvers to one file
The combined file is still small and easy to reason with
