+++ src/resolvers/index.ts
@@ -4,11 +4,8 @@ import orderMutationResolvers from './mutations/orders'
 import customerMutationResolvers from './mutations/customers'
 
 import cartQueryResolvers from './queries/cart'
-import brandQueryResolvers from './queries/brands'
-import productQueryResolvers from './queries/products'
+import legacyCatalogQueryResolvers from './queries/legacy_catalog'
 import pcmQueryResolvers from './queries/pcm'
-import categoryQueryResolvers from './queries/categories'
-import collectionQueryResolvers from './queries/collections'
 import customerQueryResolvers from './queries/customers'
 import rootQueryResolvers from './root'
 
@@ -21,12 +18,9 @@ export default {
     ...customerMutationResolvers,
   },
   Query: {
-    ...brandQueryResolvers,
     ...cartQueryResolvers,
-    ...productQueryResolvers,
     ...pcmQueryResolvers,
-    ...categoryQueryResolvers,
-    ...collectionQueryResolvers,
-    ...customerQueryResolvers
+    ...customerQueryResolvers,
+    ...legacyCatalogQueryResolvers
   },
 }
